Backend/requirements.txt
django
djangorestframework
django-cors-headers
firebase-admin
google-generativeai
python-dotenv
